Refactor I/O helpers into dedicated package modules
