+++ library/csv_utils.py
@@ -24,7 +24,7 @@ import pandas as pd
 from pandas.api import types as ptypes
 
 from .config import Config
-from .io import write_meta_yaml
+from .io.metadata import write_meta_yaml
 from .log import logger
